testing setup fixes
