+++ tox.ini
@@ -4,11 +4,10 @@ env_list =
     lint
     docs
     notebooks
-    offline-prefetch
+    doctests
     py39-upstream-doctest
     py310-doctest
     py311-lmoments
-    py312-lmoments-doctest
 labels =
     test = py39, py310-upstream-doctest, py311, notebooks_doctests, offline-prefetch
 requires =
@@ -18,10 +17,10 @@ opts = -vv
 
 [gh]
 python =
-    3.12 = py312-coverage-lmoments-doctest
-    3.11 = py311-coverage-lmoments-sbck-doctest, offline-coverage-prefetch
-    3.10 = py310-coverage-lmoments-doctest, notebooks
-    3.9 = py39-coverage-sbck-doctest, lint, docs
+    3.12 = doctests
+    3.11 = py311-coverage-lmoments-sbck, offline-coverage-prefetch
+    3.10 = py310-coverage-lmoments, notebooks
+    3.9 = py39-coverage-sbck, lint, docs
 
 [testenv:lint]
 description = Run code quality compliance tests under {basepython}
@@ -88,19 +87,11 @@ commands =
     pytest --no-cov --nbval --dist=loadscope --rootdir=tests/ --ignore=docs/notebooks/example.ipynb docs/notebooks
 commands_post =
 
-[testenv:offline{-prefetch,}{-coverage,}]
-description = Run tests with pytest under {basepython}, preventing socket connections (except for unix sockets for async support)
-commands:
-    python -c 'print("Running offline tests with positional arguments: --disable-socket --allow-unix-socket --m \"not requires_internet\"")'
-    python -c 'print("These can be overwritten with: tox -e offline -- -m \"some other marker statement\"")'
-    pytest --disable-socket --allow-unix-socket {posargs:-m 'not requires_internet'}
-allowlist_externals =
-    xclim
-
-[testenv:doctest]
+[testenv:doctests]
 description = Run doctests with pytest under {basepython}
 commands =
     python -c 'from xclim.testing import helpers; helpers.test_doctests()'
+commands_post =
 
 [testenv]
 description = Run tests with pytest under {basepython}
@@ -135,7 +126,11 @@ commands_pre =
     xclim --help
     prefetch: xclim prefetch_testing_data
 commands =
-    pytest {posargs}
+    !offline: pytest {posargs}
+    ; Prevent socket connections (except for unix sockets for async support)
+    offline: python -c 'print("Running offline tests with positional arguments: --disable-socket --allow-unix-socket --m \"not requires_internet\"")'
+    offline: python -c 'print("These can be overwritten with: tox -e offline -- -m \"some other marker statement\"")'
+    offline: pytest --disable-socket --allow-unix-socket {posargs:-m 'not requires_internet'}
 commands_post =
     coverage: - coveralls
 allowlist_externals =
